Closes #236
- Sublassing errors in case of expected behaviour
- When receiving a new message that was using a previously unknown
identity key, we search for any other messages that are encrypted with
the new accepted key and decrypt those too.
- Addressing some of the warnings we had before.

+++ Signal/src/phone/signaling/ResponderSessionDescriptor.h
@@ -11,13 +11,13 @@
  */
 @interface ResponderSessionDescriptor : NSObject
 
-@property (nonatomic,readonly) int32_t interopVersion;
+@property (nonatomic,readonly) NSUInteger interopVersion;
 @property (nonatomic,readonly) in_port_t relayUdpPort;
 @property (nonatomic,readonly) int64_t sessionId;
 @property (nonatomic,readonly) NSString* relayServerName;
 @property (nonatomic,readonly) PhoneNumber* initiatorNumber;
 
-+(ResponderSessionDescriptor*)responderSessionDescriptorWithInteropVersion:(int32_t)interopVersion
++(ResponderSessionDescriptor*)responderSessionDescriptorWithInteropVersion:(NSUInteger)interopVersion
                                                            andRelayUdpPort:(in_port_t)relayUdpPort
                                                               andSessionId:(int64_t)sessionId
                                                         andRelayServerName:(NSString*)relayServerName
